+++ architecture/api.py
@@ -4,10 +4,12 @@ import json
 from decouple import config
 from hashlib import sha512
 from certifier import Certifier
-from CAKEClient import CAKEClient
-from CAKEDataOwner import CAKEDataOwner
+from client import CAKEClient
+from data_owner import CAKEDataOwner
+from flask_cors import CORS, cross_origin
 
 app = Flask(__name__)
+CORS(app)
 
 def getClientArgs(request):
     """ Read the arguments from the client request
@@ -28,14 +30,6 @@ def getClientArgs(request):
     reader_address = request.json.get('reader_address')
     message_id = request.json.get('message_id')
     slice_id = request.json.get('slice_id')
-
-    '''
-    print("Reader_address is: " + reader_address)
-    print("Message_id is: " + message_id)
-    if slice_id is not None:
-        print("Slice_id is: " + slice_id)
-    print("Process_id is: " + str(process_id))
-    '''
     return reader_address, message_id, slice_id, process_id
 
 @app.route('/')
@@ -123,24 +117,7 @@ def accessData():
     #client.disconnect()   
 
     return "Data accessed", 200
-'''
-# This is a full request, it does handshake, generate key and access data
-@app.route('/client/fullrequest/' , methods=['GET', 'POST'])
-def client_fullrequest():
-    reader_address, message_id, slice_id, process_id = getClientArgs(request)
-    if reader_address == '' or message_id == '' or slice_id == '':
-        return "Missing parameters" , 400
-    client = CAKEClient(message_id=message_id, reader_address=reader_address, slice_id=slice_id, process_instance_id= process_id)
-    client.handshake()
-    print("Handshake launched")
-    client = CAKEClient(message_id=message_id, reader_address=reader_address, slice_id=slice_id, process_instance_id= process_id)
-    client.generate_key()
-    print("Key generation launched")    
-    client = CAKEClient(message_id=message_id, reader_address=reader_address, slice_id=slice_id, process_instance_id= process_id)
-    client.access_data()
-    print("Data access launched")
-    return "Handshake completed"
-'''
+
 
 ##### Request from Data Owner to SDM Server #####
 
@@ -203,33 +180,6 @@ def cipher():
     data_owner = CAKEDataOwner(process_instance_id=request.json.get('process_id'))
     data_owner.cipher_data(message, entries_string, policy_string)
     return "Cipher completed"
-
-'''
-@app.route('/dataOwner/fullrequest/', methods=['POST'])
-def dataowner_fullrequest():
-    message = request.json.get('message')
-    if len(message) == 0:
-        return "Missing parameters" , 400
-    entries = request.json.get('entries')
-    policy = request.json.get('policy')
-    if len(entries) == 0:
-        return "Missing parameters" , 400
-    if len(policy) == 0:
-        return "Missing parameters" , 400
-    
-    #TODO: Check if it is mandatory
-    if len(entries) != len(policy):
-        return "Entries and policy legth doesn't match" , 400   
-
-    entries_string = '###'.join(str(x) for x in entries)
-    policy_string = '###'.join(str(x) for x in policy)
-    message = json.dumps(message)
-    data_owner = CAKEDataOwner(process_instance_id=request.json.get('process_id'))
-    data_owner.handshake()
-    data_owner = CAKEDataOwner(process_instance_id=request.json.get('process_id'))
-    data_owner.cipher_data(message, entries_string, policy_string)
-    return "Full request completed"
-'''
     
 @app.route('/certification/', methods=['POST'])
 def certification():
