+++ architecture/certification.sh
@@ -1,14 +1,18 @@
 # Read public key of manufacter and suppliers
 set -e
-python3 reader_public_key.py --reader 'MANUFACTURER'
-echo "✅ Readed public key of MANUFACTURER"
-python3 reader_public_key.py --reader 'SUPPLIER1'
-echo "✅ Readed public key of SUPPLIER1"
-python3 reader_public_key.py --reader 'SUPPLIER2'
-echo "✅ Readed public key of SUPPLIER2"
 
-python3 skm_public_key.py
-echo "✅ Readed public key of skm"
+# READ PUBLIC KEY OF MANUFACTURER AND SUPPLIERS
+python3 certifier.py -o 0 --reader 'MANUFACTURER'
+echo "✅ Read public key of MANUFACTURER"
+python3  certifier.py -o 0  --reader 'SUPPLIER1'
+echo "✅ Read public key of SUPPLIER1"
+python3  certifier.py -o 0  --reader 'SUPPLIER2'
+echo "✅ Read public key of SUPPLIER2"
 
-python3 attribute_certifier.py 
+ # READ PUBLIC KEY OF SKM
+python3  certifier.py -o 1
+echo "✅ Read public key of skm"
+
+# CERTIFY ATTRIBUTES
+python3  certifier.py -o 2
 echo "✅ Attribute certifier done"
