+++ architecture/data_owner.py
@@ -5,121 +5,129 @@ import json
 from decouple import config
 import sqlite3
 import argparse
+from connector import Connector
 
-process_instance_id = config('PROCESS_INSTANCE_ID')
+class CAKEDataOwner(Connector):
+    """Tools to communicate from the API to the CAKE SDM server
 
-HEADER = 64
-PORT = int(config('SDM_PORT'))
-FORMAT = 'utf-8'
-server_sni_hostname = config('SERVER_SNI_HOSTNAME')
-DISCONNECT_MESSAGE = "!DISCONNECT"
-SERVER = config('SERVER')
-ADDR = (SERVER, PORT)
-server_cert = 'Keys/server.crt'
-client_cert = 'Keys/client.crt'
-client_key = 'Keys/client.key'
+    A class to manage the communication with the CAKE SDM server
 
-# Connection to SQLite3 data_owner database
-connection = sqlite3.connect('files/data_owner/data_owner.db')
-x = connection.cursor()
+    Attributes:
+        manufacturer_address (str): manufacturer address
+    """
 
-"""
-creation and connection of the secure channel using SSL protocol
-"""
+    def __init__(self, process_instance_id = config('PROCESS_INSTANCE_ID')):
+        """Initialize the CAKEDataOwner class
 
-context = ssl.create_default_context(ssl.Purpose.SERVER_AUTH, cafile=server_cert)
-context.load_cert_chain(certfile=client_cert, keyfile=client_key)
-s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-conn = context.wrap_socket(s, server_side=False, server_hostname=server_sni_hostname)
-conn.connect(ADDR)
+        Args:
+            process_instance_id (int, optional): process instance id. Defaults to config('PROCESS_INSTANCE_ID').
+        """        
+        super().__init__("files/data_owner/data_owner.db", int(config('SDM_PORT')), process_instance_id=process_instance_id)
+        self.manufacturer_address = config('ADDRESS_MANUFACTURER')
+        return
+    
+    """
+    function to handle the sending and receiving messages.
+    """
+    def send(self, msg):
+        """Send a message to the CAKE SDM server
 
-manufacturer_address = config('ADDRESS_MANUFACTURER')
+        Send a message to the CAKE SDM server and receive a response
 
+        Args:
+            msg (str): message to send
+        """
+        message = msg.encode(self.FORMAT)
+        msg_length = len(message)
+        send_length = str(msg_length).encode(self.FORMAT)
+        send_length += b' ' * (self.HEADER - len(send_length))
+        self.conn.send(send_length)
+        # print(send_length)
+        self.conn.send(message)
+        receive = self.conn.recv(6000).decode(self.FORMAT)
+        #print(receive)
+        if receive.startswith('Number to be signed: '):
+            len_initial_message = len('Number to be signed: ')
+            self.x.execute("INSERT OR IGNORE INTO handshake_number VALUES (?,?,?)",
+                    (self.process_instance_id, self.manufacturer_address, receive[len_initial_message:]))
+            self.connection.commit()
+        if receive.startswith('Here is the message_id:'):
+            self.x.execute("INSERT OR IGNORE INTO messages VALUES (?,?,?)", (self.process_instance_id, self.manufacturer_address, receive[16:]))
+            self.connection.commit()
 
-def sign_number():
-    x.execute("SELECT * FROM handshake_number WHERE process_instance=?", (process_instance_id,))
-    result = x.fetchall()
-    number_to_sign = result[0][2]
+    def handshake(self):
+        """Handshake with the CAKE SDM server"""
 
-    x.execute("SELECT * FROM rsa_private_key WHERE reader_address=?", (sender,))
-    result = x.fetchall()
-    private_key = result[0]
-    private_key_n = int(private_key[1])
-    private_key_d = int(private_key[2])
+        print("Start handshake")
+        self.send("Start handshake§" + self.manufacturer_address)
+        self.disconnect()
+        return
+    
+    def cipher_data(self, message_to_send, entries_string, policy_string):
+        """Cipher a message and set the policy
+        
+        Args:
+            message_to_send (str): message to send 
+            entries_string (str): entries converted to string
+            policy_string (str): policy converted to string"""
+        signature_sending = self.sign_number()
+        self.send("Cipher this message§" + message_to_send + '§' + entries_string + '§' + policy_string + '§' + self.manufacturer_address   + '§' + str(signature_sending))
+        self.disconnect()
+        return
+    
+    def sign_number(self):
+        """Sign a number
 
-    msg = bytes(str(number_to_sign), 'utf-8')
-    hash = int.from_bytes(sha512(msg).digest(), byteorder='big')
-    signature = pow(hash, private_key_d, private_key_n)
-    # print("Signature:", hex(signature))
-    return signature
+        Sign a number and return the signature
 
+        Returns:
+            str: signature
+        """
+        self.x.execute("SELECT * FROM handshake_number WHERE process_instance=?", (self.process_instance_id,))
+        result = self.x.fetchall()
+        number_to_sign = result[0][2]
+        return super().sign_number(number_to_sign, self.manufacturer_address)
 
-"""
-function to handle the sending and receiving messages.
-"""
+if __name__ == "__main__":
+    NO_SLICE = False
+    # f = open('files/data.json')
+    manufacturer_address = config('ADDRESS_MANUFACTURER')
+    g = open('files/data.json')
 
+    process_instance_id = config('PROCESS_INSTANCE_ID')
 
-def send(msg):
-    message = msg.encode(FORMAT)
-    msg_length = len(message)
-    send_length = str(msg_length).encode(FORMAT)
-    send_length += b' ' * (HEADER - len(send_length))
-    conn.send(send_length)
-    conn.send(message)
-    receive = conn.recv(6000).decode(FORMAT)
-    if receive.startswith('Number to be signed: '):
-        len_initial_message = len('Number to be signed: ')
-        x.execute("INSERT OR IGNORE INTO handshake_number VALUES (?,?,?)",
-                    (process_instance_id, sender, receive[len_initial_message:]))
-        connection.commit()
-    if receive.startswith('Here is the message_id:'):
-        x.execute("INSERT OR IGNORE INTO messages VALUES (?,?,?)", (process_instance_id, sender, receive[16:]))
-        connection.commit()
+    message_to_send = g.read()
 
+    # policy_string = '1604423002081035210 and (MANUFACTURER or (SUPPLIER and ELECTRONICS))'
 
-# f = open('files/data.json')
-g = open('files/data.json')
+    entries = [['ID', 'SortAs', 'GlossTerm'], ['Acronym', 'Abbrev'], ['Specs', 'Dates']]
+    entries_string = '###'.join(str(x) for x in entries)
+    #print(entries_string)
+    policy = [process_instance_id + ' and (MANUFACTURER or SUPPLIER)',
+            process_instance_id + ' and (MANUFACTURER or (SUPPLIER and ELECTRONICS))',
+            process_instance_id + ' and (MANUFACTURER or (SUPPLIER and MECHANICS))']
 
-message_to_send = g.read()
+    policy_string = '###'.join(policy)
 
-###########
-###########
-###########
-###LINES###
-###########
-###########
-###########
+    if NO_SLICE:
+        data = json.load(open('files/data.json'))
+        entries = [list(data.keys())]
+        entries_string = '###'.join(str(x) for x in entries)
+        print(entries_string)
 
-# policy_string = '1604423002081035210 and (MANUFACTURER or (SUPPLIER and ELECTRONICS))'
+        policy = process_instance_id + ' and (MANUFACTURER or (SUPPLIER and ELECTRONICS))'
+        policy = [policy]
+        policy_string = '###'.join(policy)
 
-entries = [['ID', 'SortAs', 'GlossTerm'], ['Acronym', 'Abbrev'], ['Specs', 'Dates']] 
-entries_string = '###'.join(str(x) for x in entries)
+    sender = manufacturer_address
 
-policy = [str(process_instance_id) + ' and (MANUFACTURER or SUPPLIER)',
-          str(process_instance_id) + ' and (MANUFACTURER or (SUPPLIER and ELECTRONICS))',
-          str(process_instance_id) + ' and (MANUFACTURER or (SUPPLIER and MECHANICS))']
-policy_string = '###'.join(policy)
+    parser = argparse.ArgumentParser()
+    parser.add_argument('-hs' ,'--hanshake', action='store_true')
+    parser.add_argument('-c','--cipher', action='store_true')
+    dataOwner= CAKEDataOwner()
+    args = parser.parse_args()
+    if args.hanshake:
+        dataOwner.handshake()
 
-# data = json.load(f)
-# entries = list(data.keys())
-# entries_string = '###'.join(entries)
-# print(entries_string)
-# exit()
-
-# entries_string = ''
-
-sender = manufacturer_address
-
-parser = argparse.ArgumentParser()
-parser.add_argument('-hs' ,'--hanshake', action='store_true')
-parser.add_argument('-c','--cipher', action='store_true')
-
-args = parser.parse_args()
-if args.hanshake:
-    send("Start handshake§" + sender)
-
-if args.cipher:
-    signature_sending = sign_number()
-    send("Cipher this message§" + message_to_send + '§' + entries_string + '§' + policy_string + '§' + sender + '§' + str(signature_sending))
-
-send(DISCONNECT_MESSAGE)
+    if args.cipher:
+        dataOwner.cipher_data(message_to_send, entries_string, policy_string)
