+++ architecture/data_owner.sh
@@ -1,4 +1,4 @@
 python3 data_owner.py -hs
 echo "✅ Data owner handshake done"
 python3 data_owner.py -c
-echo "✅ Data owner certification done"
+echo "✅ Data owner cipher done"
